acceptor/requirements.txt
Flask==3.1.0
google-cloud-tasks==2.19.3
google-cloud-logging==3.12.1
gunicorn==23.0.0
google-cloud-secret-manager==2.24.0